Fixed sticky keys issue and added new enemy type

+++ Game/src/BasicEnemy.java
@@ -5,7 +5,7 @@ public class BasicEnemy extends GameObject {
     private Handler handler;
 
     //constructor
-    public BasicEnemy(int x, int y, ID id, Handler handler) {
+    public BasicEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
 
         this.handler = handler;
@@ -15,7 +15,7 @@ public class BasicEnemy extends GameObject {
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int) x, (int) y, 16, 16);
     }
 
     //makes our enemy move
@@ -37,7 +37,7 @@ public class BasicEnemy extends GameObject {
     //draws our enemy
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int) x, (int) y, 16, 16);
     }
     
 }
