+++ Game/src/FastEnemy.java
@@ -4,7 +4,7 @@ public class FastEnemy extends GameObject{
     private Handler handler;
 
     //constructor
-    public FastEnemy(int x, int y, ID id, Handler handler) {
+    public FastEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
 
         this.handler = handler;
@@ -14,7 +14,7 @@ public class FastEnemy extends GameObject{
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int) x, (int) y, 16, 16);
     }
 
     //makes our enemy move
@@ -36,6 +36,6 @@ public class FastEnemy extends GameObject{
     //draws our enemy
     public void render(Graphics g) {
         g.setColor(Color.CYAN);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int) x, (int) y, 16, 16);
     }
 }
