+++ Game/src/HUD.java
@@ -2,8 +2,8 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int health = 100;
-    private int greenValue = 255;
+    public static float health = 100;
+    private float greenValue = 255;
 
     private int score = 0;
     private int level = 1;
@@ -23,8 +23,8 @@ public class HUD {
         g.setColor(Color.gray);
         g.fillRect(15, 15, 200, 32);
 
-        g.setColor(new Color(150, greenValue, 0));
-        g.fillRect(15, 15, health * 2, 32);
+        g.setColor(new Color(150, (int) greenValue, 0));
+        g.fillRect(15, 15, (int) health * 2, 32);
 
         g.setColor(Color.white);
         //puts a border around the health bar
